+++ simple_history/tests/urls.py
@@ -3,7 +3,7 @@ from __future__ import unicode_literals
 from django.conf.urls import url
 from django.contrib import admin
 
-from simple_history.tests.view import (
+from .view import (
     BucketDataRegisterRequestUserCreate,
     BucketDataRegisterRequestUserDetail,
     PollCreate,