+++ CHANGES.rst
@@ -2,6 +2,8 @@ Changes
 =======
 
 - Fixed most_recent() bug with excluded_fields (gh-561)
+- Add support for historical model permissions, including DJ2.1 view permission (gh-529)
+- Add option to disable a user's, or all user's, ability to "revert" to previous version (gh-529)
 
 
 2.7.2 (2019-04-17)